Version 6.1.0
Various minor and more major improvements on the development version over the last year,
which are now ported in this release.

1) A completely new CMakeLists file that also installs various external libraries if they
are required (boost, eigen, etc.) or requested (pastix, Umfpack, etc.) and not found in
the system.

CMakeLists.txt has been replaced by CMakeLists-template.txt to prevent git from clashing with
local, user-specific choices (just copy this template into CMakeLists.txt and modify the later).

2) A new class, "LinearSolver", can be used in your code to allow you to select the required solver
on the fly. See example of usage in HHO-full-gradient-diff.

3) Implementation of a DDR complex for the quadrot problem.

4) Creation of a vector version of HHOCore, e.g. for Stokes and other problems involving vector-valued
functions.